address_ranges {tigris}	R Documentation

Download an address range features shapefile into R

Description

Download an address range features shapefile into R

Usage

address_ranges(state, county, year = NULL, ...)

Arguments

state
The two-digit FIPS code of the state of the county you'd like to download the roads for. Can also be state name or abbreviation (case-insensitive).

county
The three-digit FIPS code of the county you'd like the roads for. Can also be a county name.

year
the data year (defaults to 2019).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sp" (the default) or "sf" to request sp or sf class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

http://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2019/TGRSHP2019_TechDoc.pdf

Other transportation functions: primary_roads(), primary_secondary_roads(), rails(), roads()

[Package tigris version 1.0 Index]