alaska_native_regional_corporations {tigris}	R Documentation

Download an Alaska Native Regional Corporation shapefile into R.

Description

From the US Census Bureau: "ANRCs are corporations created according to the Alaska Native Claims Settlement Act. They are organized under the laws of the State of Alaska as "Regional Corporations," to conduct both the for-profit and non-profit affairs of Alaska Natives within defined regions of Alaska." For more information, please see the Census technical documentation at the link provided.

Usage

alaska_native_regional_corporations(cb = FALSE, year = NULL, ...)

Arguments

cb
If cb is set to TRUE, download a generalized (1:500k) file. Defaults to FALSE (the most detailed TIGER/Line file)

year
the data year (defaults to 2017).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sp" (the default) or "sf" to request sp or sf class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

http://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2015/TGRSHP2015_TechDoc.pdf

Other native/tribal geometries functions: native_areas, tribal_block_groups, tribal_census_tracts, tribal_subdivisions_national

[Package tigris version 0.8.2 Index]