area_water {tigris}	R Documentation

Download an area water shapefile into R

Description

From the US Census Bureau: "The area hydrography shapefile contains the geometry and attributes of both perennial and intermittent area hydrography features, including ponds, lakes, oceans, swamps, glaciers, and the area covered by large streams represented as double-line drainage."

Usage

area_water(state, county, year = NULL, ...)

Arguments

state
The two-digit FIPS code of the state of the county you'd like to download the water features for. Can also be state name or abbreviation (case-insensitive).

county
The three-digit FIPS code of the county you'd like the water features for. Can also be a county name.

year
the data year (defaults to 2017).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sp" (the default) or "sf" to request sp or sf class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).).

See Also

http://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2015/TGRSHP2015_TechDoc.pdf

Other water functions: coastline, linear_water

Examples

## Not run: 
library(tigris)
library(sp)

dallas_water <- area_water("TX", "Dallas")

plot(dallas_water)

## End(Not run)

[Package tigris version 0.8.2 Index]