call_geolocator {tigris}	R Documentation

Call geolocator for one address

Description

Call geolocator for one address

Usage

call_geolocator(street, city, state, zip = NA)

Arguments

street
A character string indicating a street name and number

city
A character string indicating a city

state
A two-digit character string with a state postal code

zip
A five-digit character string with a postal zip code. Optional parameter.

Value

A character string representing the Census block of the supplied address.

importFrom utils URLencode importFrom httr GET stop_for_status

[Package tigris version 1.0 Index]