call_geolocator_latlon {tigris}	R Documentation

Call geolocator for one address with lat/lon, adds option to set benchmark and vintage if not provided defualts to most recent.

Description

Call geolocator for one address with lat/lon, adds option to set benchmark and vintage if not provided defualts to most recent.

Usage

call_geolocator_latlon(lat, lon, benchmark, vintage)

Arguments

lat
A numeric value

lon
A numeric value

benchmark
time period when a snapshot of address ranges was taken

vintage
census or survey that the address range relates to

Value

A character string representing the Census block of the supplied lat/lon.

Author(s)

Josie Kressner, josie@transportfoundry.com

Mark Richards, Mark.Richards.002@gmail.com

[Package tigris version 0.8.2 Index]