coastline {tigris}	R Documentation

Download a shapefile of the US coastline into R

Description

Download a shapefile of the US coastline into R

Usage

coastline(year = NULL, ...)

Arguments

year
The year of the dataset (defaults to 2019)

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sp" (the default) or "sf" to request sp or sf class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).).

See Also

Other water functions: area_water(), linear_water()

[Package tigris version 1.0 Index]