combined_statistical_areas {tigris}	R Documentation

Download a combined statistical areas shapefile into R

Description

Combined statistical areas are "two or more adjacent CBSAs that have significant employment interchanges." In turn, CSAs are composed of multiple metropolitan and/or micropolitan areas, and should not be compared with individual core-based statistical areas.

Usage

combined_statistical_areas(cb = FALSE, resolution = "500k",
  year = NULL, ...)

Arguments

cb
If cb is set to TRUE, download a generalized (1:500k) cartographic boundary file. Defaults to FALSE (the most detailed TIGER/Line file).

resolution
The resolution of the cartographic boundary file (if cb == TRUE). Defaults to '500k'; options include '5m' (1:5 million) and '20m' (1:20 million).

year
the data year (defaults to 2017).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sp" (the default) or "sf" to request sp or sf class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

http://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2015/TGRSHP2015_TechDoc.pdf

Other metro area functions: core_based_statistical_areas, metro_divisions, new_england, urban_areas

[Package tigris version 0.8.2 Index]