core_based_statistical_areas {tigris}	R Documentation

Download a core-based statistical area shapefile into R

Description

Core-based statistical areas include both metropolitan areas and micropolitan areas. The US Census Bureau defines these areas as follows: "A metro area contains a core urban area of 50,000 or more population, and a micro area contains an urban core of at least 10,000 (but less than 50,000) population. Each metro or micro area consists of one or more counties and includes the counties containing the core urban area, as well as any adjacent counties that have a high degree of social and economic integration (as measured by commuting to work) with the urban core." Please see the link provided for more information

Usage

core_based_statistical_areas(cb = FALSE, resolution = "500k",
  year = NULL, ...)

Arguments

cb
If cb is set to TRUE, download a generalized (1:500k) cartographic boundary file. Defaults to FALSE (the most detailed TIGER/Line file).

resolution
The resolution of the cartographic boundary file (if cb == TRUE). Defaults to '500k'; options include '5m' (1:5 million) and '20m' (1:20 million).

year
the data year (defaults to 2017).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sp" (the default) or "sf" to request sp or sf class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

http://www.census.gov/population/metro/

Other metro area functions: combined_statistical_areas, metro_divisions, new_england, urban_areas

[Package tigris version 0.8.2 Index]