county_subdivisions {tigris}	R Documentation

Download a county subdivision shapefile into R

Description

From the US Census Bureau (see link for source, and more information): "All counties and statistically equivalent entities consist of one or more geographic units that the Bureau of the Census recognizes as county subdivisions. The two major types of county subdivisions are minor civil divisions(MCDs) and census county divisions (CCDs). A State has either MCDs or their statistical equivalents, or CCDs; it cannot contain both."

Usage

county_subdivisions(state, county = NULL, cb = FALSE, year = NULL, ...)

Arguments

state
The two-digit FIPS code (string) of the state you want. Can also be state name or state abbreviation.

county
The three-digit FIPS code (string) of the county you'd like to subset for, or a vector of FIPS codes if you desire multiple counties. Can also be a county name or vector of names.

cb
If cb is set to TRUE, download a generalized (1:500k) file. Defaults to FALSE (the most detailed TIGER/Line file)

year
the data year (defaults to 2019).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sf" (the default) or "sp" to request sf or sp class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

https://www2.census.gov/geo/pdfs/reference/GARM/Ch8GARM.pdf

Other general area functions: block_groups(), blocks(), counties(), places(), pumas(), school_districts(), states(), tracts(), zctas()

Examples

## Not run: 
library(tigris)

or <- county_subdivisions('Oregon', c('Linn', 'Benton'))

plot(or$geometry)

## End(Not run)

[Package tigris version 1.0 Index]