filter_place {tigris}	R Documentation

Filter a places Spatial object for only those places matching the contents of the place vector.

Description

Filter a places Spatial object for only those places matching the contents of the place vector.

Usage

filter_place(places, place)

Arguments

places
object returned from a call to places

place
a vector of full place names. The function performs the comparison in a case-insensitive manner.

Examples

## Not run: 
places("Maine") %>% filter_place("berwick")

## End(Not run)

[Package tigris version 0.8.2 Index]