filter_state {tigris}	R Documentation

Filter a states Spatial object for only those states matching the contents of the state vector.

Description

Filter a states Spatial object for only those states matching the contents of the state vector.

Usage

filter_state(states, state)

Arguments

states
object returned from a call to states

state
a vector of full state names. The function performs the comparison in a case-insensitive manner.

Examples

## Not run: 
states() %>% filter_state("south")

## End(Not run)

[Package tigris version 1.0 Index]