fips_codes {tigris}	R Documentation

Dataset with FIPS codes for US states and counties

Description

Built-in dataset for use with the lookup_code function. To access the data directly, issue the command data(fips_codes).

county: County name, title-case

county_code: County code. (3-digit, 0-padded, character)

state: Upper-case abbreviation of state

state_code: State FIPS code (2-digit, 0-padded, character)

state_name: Title-case name of state

Usage

data(fips_codes)

Format

A data frame with 3,236 rows and 5 variables

Details

Dataset with FIPS codes for US states and counties

Built-in dataset for use with the lookup_code function. To access the data directly, issue the command data(fips_codes).

Note

Last updated 2015-07-01

[Package tigris version 1.0 Index]