list_places {tigris}	R Documentation

Return a list of all the places in a places object

Description

Return a list of all the places in a places object

Usage

list_places(places, sorted = TRUE)

Arguments

places
object returned from a call to places

sorted
return the list sorted or in the order found in the shapefile?

Examples

## Not run: 
places("Maine") %>% list_places()

## End(Not run)

[Package tigris version 1.0 Index]