list_states {tigris}	R Documentation

Return a list of all the states in a state object

Description

Return a list of all the states in a state object

Usage

list_states(states, sorted = TRUE)

Arguments

states
object returned from a call to state

sorted
return the list sorted or in the order found in the shapefile?

Examples

## Not run: 
states() %>% list_states()

## End(Not run)

[Package tigris version 0.8.2 Index]