places {tigris}	R Documentation

Download a Census-designated places shapefile into R

Description

Census Designated Places (CDPs) are the statistical counterparts of incorporated places, and are delineated to provide data for settled concentrations of population that are identifiable by name but are not legally incorporated under the laws of the state in which they are located.

Usage

places(state, cb = FALSE, year = NULL, ...)

Arguments

state
The two-digit FIPS code (string) of the state you want. Can also be state name or state abbreviation.

cb
If cb is set to TRUE, download a generalized (1:500k) cartographic boundary file. Defaults to FALSE (the most detailed TIGER/Line file).

year
the data year (defaults to 2019).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sf" (the default) or "sp" to request sf or sp class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

Details

The boundaries usually are defined in cooperation with local or tribal officials and generally updated prior to each decennial census.

These boundaries, which usually coincide with visible features or the boundary of an adjacent incorporated place or another legal entity boundary, have no legal status, nor do these places have officials elected to serve traditional municipal functions.

CDP boundaries may change from one decennial census to the next with changes in the settlement pattern; a CDP with the same name as in an earlier census does not necessarily have the same boundary.

CDPs must be contained within a single state and may not extend into an incorporated place.

There are no population size requirements for CDPs.

See Also

https://www2.census.gov/geo/pdfs/reference/GARM/Ch9GARM.pdf

Other general area functions: block_groups(), blocks(), counties(), county_subdivisions(), pumas(), school_districts(), states(), tracts(), zctas()

[Package tigris version 1.0 Index]