primary_secondary_roads {tigris}	R Documentation

Download a primary & secondary roads shapefile into R

Description

From the Census Bureau: "Primary roads are generally divided, limited-access highways within the Federal interstate highway system or under state management. These highways are distinguished by the presence of interchanges and are accessible by ramps and may include some toll highways. Secondary roads are main arteries, usually in the U.S. highway, state highway, or county highway system. These roads have one or more lanes of traffic in each direction, may or may not be divided, and usually have at-grade intersections with many other roads and driveways.

Usage

primary_secondary_roads(state, year = NULL, ...)

Arguments

state
The two-digit FIPS code of the state of the county you'd like to download the roads for. Can also be state name or abbreviation (case-insensitive).

year
the data year (defaults to 2019).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sf" (the default) or "sp" to request sf or sp class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

http://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2019/TGRSHP2019_TechDoc.pdf

Other transportation functions: address_ranges(), primary_roads(), rails(), roads()

Examples

## Not run: 
library(tigris)

rds <- primary_secondary_roads()

plot(rds$geometry)

## End(Not run)

[Package tigris version 1.0 Index]