pumas {tigris}	R Documentation

Download a Public Use Microdata Area (PUMA) shapefile into R

Description

Public use microdata areas (PUMAs) are decennial census areas that have been defined for the tabulation and dissemination of Public Use Microdata Sample (PUMS) data, American Community Survey (ACS) data, and ACS period estimates. For the 2010 Census, the State Data Centers (SDCs) in each state, the District of Columbia, and the Commonwealth of Puerto Rico were given the opportunity to delineate PUMAs within their state or statistically equivalent entity. All PUMAs must nest within states and have a minimum population threshold of 100,000 persons. 2010 PUMAs were built on census tracts and cover the entirety of the United States, Puerto Rico, Guam, and the U.S. Virgin Islands. Because they do not meet the minimum population requirement, the Commonwealth of the Northern Mariana Islands and American Samoa do not contain any 2010 PUMAs.

Usage

pumas(state, cb = FALSE, year = NULL, ...)

Arguments

state
The two-digit FIPS code (string) of the state you want. Can also be state name or state abbreviation.

cb
If cb is set to TRUE, download a generalized (1:500k) states file. Defaults to FALSE (the most detailed TIGER/Line file)

year
the data year (defaults to 2019).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sf" (the default) or "sp" to request sf or sp class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

https://www.census.gov/programs-surveys/geography/guidance/geo-areas/pumas.html

Other general area functions: block_groups(), blocks(), counties(), county_subdivisions(), places(), school_districts(), states(), tracts(), zctas()

Examples

## Not run: 
library(tigris)

us_states <- unique(fips_codes$state)[1:51]

continental_states <- us_states[!us_states %in% c("AK", "HI")]
pumas_list <- lapply(continental_states, function(x) {
  pumas(state = x, cb = TRUE)
  })

us_pumas <- rbind_tigris(pumas_list)

plot(us_pumas$geometry)

## End(Not run)

[Package tigris version 1.0 Index]