regions {tigris}	R Documentation

Download a US regions cartographic boundary shapefile into R

Description

Download a US regions cartographic boundary shapefile into R

Usage

regions(resolution = "500k", year = NULL, ...)

Arguments

resolution
The resolution of the cartographic boundary file. Defaults to '500k'; options include '5m' (1:5 million) and '20m' (1:20 million).

year
the data year (defaults to 2017).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sp" (the default) or "sf" to request sp or sf class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

Other national cartographic boundary functions: divisions, nation

Examples

## Not run: 
library(tigris)
library(leaflet)

us_regions <- regions(resolution = '20m')

leaflet(us_regions) %>%
   addTiles() %>%
   addPolygons()

## End(Not run)

[Package tigris version 0.8.2 Index]