school_districts {tigris}	R Documentation

Download a school district shapefile into R

Description

From the US Census Bureau (see link for source): School Districts are single-purpose administrative units within which local officials provide public educational services for the area's residents. The Census Bureau obtains school district boundaries, names, local education agency codes, grade ranges, and school district levels biennially from state education officials. The Census Bureau collects this information for the primary purpose of providing the U.S. Department of Education with annual estimates of the number of children in poverty within each school district, county, and state. This information serves as the basis for the Department of Education to determine the annual allocation of Title I funding to states and school districts.

Usage

school_districts(state, type = "unified", year = NULL, ...)

Arguments

state
The two-digit FIPS code (string) of the state you want. Can also be state name or state abbreviation.

type
Specify whether you want to return a unified school district (the default, 'unified'), an elementary school district ('elementary'), or a secondary school district ('secondary'). Please note: elementary and secondary school districts do not exist in all states

year
the data year; defaults to 2017

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sp" (the default) or "sf" to request sp or sf class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

Details

The Census Bureau creates pseudo-unified school districts for areas in which unified school districts do not exist. Additionally, elementary and secondary school districts do not exist in all states. Please see the link for more information on how the Census Bureau creates the school district shapefiles.

See Also

http://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2015/TGRSHP2015_TechDoc.pdf

Other general area functions: block_groups, blocks, counties, county_subdivisions, places, pumas, states, tracts, zctas

Examples

## Not run: 
library(tigris)
library(leaflet)

schools <- school_districts("Maine")

leaflet(schools) %>%
  addProviderTiles("CartoDB.Positron") %>%
  addPolygons(fillColor = "white",
              color = "black",
              weight = 0.5)

## End(Not run)

[Package tigris version 0.8.2 Index]