state_legislative_districts {tigris}	R Documentation

Download a state legislative districts shapefile into R - upper or lower

Description

This function allows you to download boundaries for state legislatures into R. Generally, state legislatures are comprised of an "upper" house, which is typically referred to as the Senate, and a "lower" house, which is often (but not exclusively) referred to as the House. The exception is Nebraska, which has a unicameral state legislature.

Usage

state_legislative_districts(
  state,
  house = "upper",
  cb = FALSE,
  year = NULL,
  ...
)

Arguments

state
The two-digit FIPS code (string) of the state. Can also be state name or abbreviation (case-insensitive)

house
Specify here whether you want boundaries for the upper or lower house. Defaults to upper.

cb
If cb is set to TRUE, download a generalized (1:500k) cartographic boundary file. Defaults to FALSE (the most detailed TIGER/Line file).

year
the data year (defaults to 2017).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sf" (the default) or "sp" to request sf or sp class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

https://www.census.gov/programs-surveys/geography/guidance/geo-areas/state-legis-dist.html

Other legislative district functions: congressional_districts(), voting_districts()

Examples

## Not run: 
library(tigris)
library(leaflet)

leg <- state_legislative_districts("Maine", "lower", cb = TRUE)

leaflet(leg) %>%
  addProviderTiles("CartoDB.Positron") %>%
  addPolygons(fillColor = "white",
              color = "black",
              weight = 0.5)

## End(Not run)

[Package tigris version 1.0 Index]