states {tigris}	R Documentation

Download shapefile for all states into R

Description

States and Equivalent Entities are the primary governmental divisions of the United States. In addition to the 50 states, the Census Bureau treats the District of Columbia, Puerto Rico, American Samoa, the Commonwealth of the Northern Mariana Islands, Guam, and the U.S. Virgin Islands as the statistical equivalents of states for the purpose of data presentation.

Usage

states(cb = FALSE, resolution = "500k", year = NULL, ...)

Arguments

cb
If cb is set to TRUE, download a generalized (1:500k) states file. Defaults to FALSE (the most detailed TIGER/Line file)

resolution
The resolution of the cartographic boundary file (if cb == TRUE). Defaults to '500k'; options include '5m' (1:5 million) and '20m' (1:20 million).

year
the year of the data download (defaults to 2017)

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sp" (the default) or "sf" to request sp or sf class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

https://www.census.gov/geo/reference/gtc/gtc_state.html

Other general area functions: block_groups, blocks, counties, county_subdivisions, places, pumas, school_districts, tracts, zctas

Examples

## Not run: 
library(tigris)
library(leaflet)

states <- states(cb = TRUE)

leaflet(states) %>%
  addProviderTiles("CartoDB.Positron") %>%
  addPolygons(fillColor = "white",
              color = "black",
              weight = 0.5) %>%
  setView(-98.5795, 39.8282, zoom=3)

## End(Not run)

[Package tigris version 0.8.2 Index]