tigris {tigris}	R Documentation

Download and use U.S. Census TIGER shapefiles in R

Description

Use option tigris_use_cache to tell tigris to cache Census shapefile downloads. This is TRUE by default. e.g. options(tigris_use_cache=TRUE)

Details

Use option tigris_refresh to force a refresh of cached tigris Shapefiles. e.g. options(tigris_refresh=TRUE)

Use option tigris_year to change the year for which you'd like to download data. e.g. options(tigris_year = 2013). The default year for the package is 2015, the most recent year for which data are available.

Note

Four options control behavior of various tigris functions. See Description for more information.

Author(s)

Kyle Walker

Kyle Walker (@kyle_e_walker)

[Package tigris version 0.8.2 Index]