tigris-exports {tigris}	R Documentation

tigris exported operators

Description

The following functions are imported and then re-exported from the tigris package to enable use of the magrittr pipe operator and the sp plot method without any additional library calls

[Package tigris version 0.8.2 Index]