tigris_type {tigris}	R Documentation

Get the type of tigris object obj is

Description

Get the type of tigris object obj is

Usage

tigris_type(obj)

Arguments

obj
R object to test

Value

character vector containing the tigris type of obj or NA if obj is not a code tigris object

[Package tigris version 1.0 Index]