tribal_subdivisions_national {tigris}	R Documentation

Download an American Indian Tribal Subdivision National shapefile into R.

Description

Definition from the US Census Bureau: "American Indian Tribal Subdivisions (AITS) are legally defined administrative subdivisions of federally recognized American Indian reservations and/or off-reservation trust lands or Oklahoma tribal statistical areas (OTSAs)." For more information, please see the link provided.

Usage

tribal_subdivisions_national(year = NULL, ...)

Arguments

year
The year for which you'd like to download data (defaults to 2019).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sf" (the default) or "sp" to request sf or sp class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

https://www2.census.gov/geo/pdfs/reference/GARM/Ch5GARM.pdf

Other native/tribal geometries functions: alaska_native_regional_corporations(), native_areas(), tribal_block_groups(), tribal_census_tracts()

Examples

## Not run: 
library(tigris)
library(leaflet)

trib <- tribal_subdivisions_national()
leaflet(trib) %>%
  addProviderTiles("CartoDB.Positron") %>%
  addPolygons(fillColor = "white",
              color = "black",
              weight = 0.5)

## End(Not run)

[Package tigris version 1.0 Index]