urban_areas {tigris}	R Documentation

Download an urban areas shapefile into R

Description

Urban areas include both "urbanized areas," which are densely developed areas with a population of at least 50,000, and "urban clusters," which have a population of greater than 2,500 but less than 50,000. For more information, please see the link provided.

Usage

urban_areas(cb = FALSE, year = NULL, ...)

Arguments

cb
If cb is set to TRUE, download a generalized (1:500k) cartographic boundary file. Defaults to FALSE (the most detailed TIGER/Line file).

year
the data year (defaults to 2017).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sp" (the default) or "sf" to request sp or sf class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

http://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2015/TGRSHP2015_TechDoc.pdf

Other metro area functions: combined_statistical_areas, core_based_statistical_areas, metro_divisions, new_england

[Package tigris version 0.8.2 Index]