zctas {tigris}	R Documentation

Download a Zip Code Tabulation Area (ZCTA) shapefile into R

Description

ZIP Code Tabulation Areas (ZCTAs) are generalized areal representations of United States Postal Service (USPS) ZIP Code service areas. Please see the link provided for information on how the Census Bureau creates ZCTAs, and for important information on the differences between ZCTAs and ZIP Codes.

Usage

zctas(cb = FALSE, starts_with = NULL, year = NULL, state = NULL, ...)

Arguments

cb
If cb is set to TRUE, download a generalized (1:500k) ZCTA file. Defaults to FALSE (the most detailed TIGER/Line file). A warning: the detailed TIGER/Line ZCTA file is massive (around 502MB unzipped), and the generalized version is also large (64MB zipped). Be prepared for this especially if you have a slower internet connection.

starts_with
Character vector specifying the beginning digits of the ZCTAs you want to return. For example, supplying the argument starts_with = c("75", "76") will return only those ZCTAs that begin with 75 or 76. Defaults to NULL, which will return all ZCTAs in the US.

year
the data year (defaults to 2019).

state
the state for which you are requesting data; only available for 2000 and 2010

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sf" (the default) or "sp" to request sf or sp class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

https://www.census.gov/programs-surveys/geography/guidance/geo-areas/zctas.html

Other general area functions: block_groups(), blocks(), counties(), county_subdivisions(), places(), pumas(), school_districts(), states(), tracts()

Examples

## Not run: 
# Example: get ZCTAs that intersect the Memphis, TN urbanized area

library(tigris)
zcta1 <- zctas(cb = TRUE, starts_with = c("37", "38", "72"))

uas <- urban_areas()

memphis_ua <- uas[grep("Memphis", uas$NAME10), ]

mem_zcta <- zcta1[memphis_ua, ]

plot(mem_zcta$geometry)

## End(Not run)

[Package tigris version 1.0 Index]