congressional_districts {tigris}	R Documentation

Download a congressional districts shapefile for the 114th Congress into R

Description

Description from the US Census Bureau (see link for source): The 2017 TIGER/Line Shapefiles contain the 115th Congressional Districts. All congressional districts appearing in the 2017 TIGER/Line Shapefiles reflect the information provided to the Census Bureau by the states by May 1, 2016. The 115th Congressional District shapefile contains the areas in effect from January 2017 to 2019.

Usage

congressional_districts(cb = FALSE, resolution = "500k", year = NULL,
  ...)

Arguments

cb
If cb is set to TRUE, download a generalized (1:500k) cartographic boundary file. Defaults to FALSE (the most detailed TIGER/Line file).

resolution
The resolution of the cartographic boundary file (if cb == TRUE). Defaults to '500k'; options include '5m' (1:5 million) and '20m' (1:20 million).

year
the data year (defaults to 2017). To get boundaries for the 115th congress, set year = 2017.

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sp" (the default) or "sf" to request sp or sf class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

Details

Congressional districts are the 435 areas from which people are elected to the U.S. House of Representatives and the five areas with nonvoting delegates from state equivalents. After the apportionment of congressional seats among the states based on decennial census population counts, each state is responsible for establishing the boundaries of the congressional districts for the purpose of electing representatives. Each congressional district is to be as equal in population to all other congressional districts in a state as practicable.

See Also

http://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2017/TGRSHP2017_TechDoc.pdf

Other legislative district functions: state_legislative_districts, voting_districts

Examples

## Not run: 
library(tigris)
library(leaflet)

cd114 <- congressional_districts(cb = TRUE, resolution = '20m')

leaflet(cd114) %>%
   addTiles() %>%
   addPolygons()

## End(Not run)

[Package tigris version 0.8.2 Index]