county_subdivisions {tigris}	R Documentation

Download a county subdivision shapefile into R

Description

From the US Census Bureau (see link for source, and more information): "County subdivisions are the primary divisions of counties and their equivalent entities for the reporting of decennial census data. They include census county divisions, census subareas, minor civil divisions, and unorganized territories. They may represent legal or statistical entities. The 2015 TIGER/Line Shapefiles contain a 5-character FIPS code field for county subdivisions and an 8-character National Standards (GNIS) code."

Usage

county_subdivisions(state, county = NULL, cb = FALSE, year = NULL,
  ...)

Arguments

state
The two-digit FIPS code (string) of the state you want. Can also be state name or state abbreviation.

county
The three-digit FIPS code (string) of the county you'd like to subset for, or a vector of FIPS codes if you desire multiple counties. Can also be a county name or vector of names.

cb
If cb is set to TRUE, download a generalized (1:500k) file. Defaults to FALSE (the most detailed TIGER/Line file)

year
the data year (defaults to 2017).

...
arguments to be passed to the underlying 'load_tiger' function, which is not exported. Options include class, which can be set to "sp" (the default) or "sf" to request sp or sf class objects, and refresh, which specifies whether or not to re-download shapefiles (defaults to FALSE).

See Also

http://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2015/TGRSHP2015_TechDoc.pdf

Other general area functions: block_groups, blocks, counties, places, pumas, school_districts, states, tracts, zctas

Examples

## Not run: 
library(tigris)

or <- county_subdivisions('Oregon', c('Linn', 'Benton'))

plot(or)

## End(Not run)

[Package tigris version 0.8.2 Index]